EE screen adjustments
1.       On the “can’t apply” scenario

Change verbiage from “You are unable to apply for this benefit” to say,
“you are ineligible for this benefit,  if you have questions please
contact HR benefits”

2.       On all scenarios including “has draft”  and after , the user
should be able to click on a link and see the submitted application
in read only format.

This should be available for the current and future academic year

3.       On the Approved Scenario, the user needs to be  see the
Award information and the Child detail information so they understand
what has been approved

Need to re-work the Current Award information to be cleaner.
Also need to add academic year for the scenarios

Link to read only view of what the administrator has for a view